feat: create unit tests for checking that heroes can equip valid weapons and armors
Create unit tests for checking that a valid weapon and piece of armor can be equipped.

Create an Equip() method for equipping heroes with a weapon.
Overload the Equip() method with another one that takes in an armor and a slot.

Create unit tests for checking that equipping an invalid weapon throws a custom exception.

Create a custom exception for equipping a weapon or armor of invalid type and an exception for equipping a weapon or armor with too high level requirement.

Modify the Equip() methods to throw exceptions if invalid weapons and armor are being equipped.