feat: create unit tests for checking that total attributes are calculated properly + refactor tests
Create unit tests for checking the calculation of total attribtues with no equipment, one piece of armor, two pieces of armor, and a replaced piece of armor.

Create the CalculateTotalAttributes() method in the Hero class, as the implementation is the same for all types of heroes.

Refactor test files and make a new test class for each of the hero types.